* Move unit tests to a dedicated folder and update grunt
