+++ grunt-config.js
@@ -22,6 +22,6 @@ module.exports = {
     },
 
     reports_dir : 'reports',
-    unit_tests : 'tests/*.js',
+    unit_tests : 'tests/unit/*.js',
     component_tests : 'tests/component/*.js'
 };
