+++ tests/unit/promise-m3u8-spec.js
@@ -29,16 +29,16 @@ describe('Promise m3u8 spec', function() {
             customizeMocks(mocks);
         }
 
-        var parser = proxyquire('../lib/promise-m3u8', mocks);
+        var parser = proxyquire('../../lib/promise-m3u8', mocks);
         return parser;
     };
 
     it('should correctly read an M3U8', function(done)
     {
-        var m3u8Parser = require('../lib/promise-m3u8');
-        var expectedManifest = fs.readFileSync(path.join(__dirname, '/resources/simpleManifest.m3u8'), 'utf8');
+        var m3u8Parser = require('../../lib/promise-m3u8');
+        var expectedManifest = fs.readFileSync(path.join(__dirname, '/../resources/simpleManifest.m3u8'), 'utf8');
 
-        m3u8Parser.parseM3U8(path.join(__dirname, '/resources/simpleManifest.m3u8')).then(function(manifest){
+        m3u8Parser.parseM3U8(path.join(__dirname, '/../resources/simpleManifest.m3u8')).then(function(manifest){
             expect(manifest.toString()).to.eql(expectedManifest.replace(/[\r]/g, ''));
         }).done(function(){
             done();
@@ -47,8 +47,8 @@ describe('Promise m3u8 spec', function() {
 
     it('should correctly read from a stream', function(done)
     {
-        var m3u8Parser = require('../lib/promise-m3u8');
-        var expectedManifest = fs.readFileSync(path.join(__dirname, '/resources/simpleManifest.m3u8'), 'utf8');
+        var m3u8Parser = require('../../lib/promise-m3u8');
+        var expectedManifest = fs.readFileSync(path.join(__dirname, '/../resources/simpleManifest.m3u8'), 'utf8');
 
         var Readable = require('stream').Readable;
         var s = new Readable();
